Refactor to move code from pop_suite to pop_suite_txn
* This commit only moves code around so that as much of the per test
  transaction logic is in the pop_suite_txn.go file
